Fixes bug where scroll was prevented in touch devices. Bumps to 1.3.
In the process, gets rid of the dependency on using jQuery.data to count setups, instead binding a flag setting handler to movestart whenever any move events are bound, which tells the default movestart handler to prevent the default of touchmove. Man, touch events are tricky.

Also, movestart no longer passes extra data arguments that were only for internal use to handlers. This was a potential bug, although no-one has bumped into it yet. Or they didn't report it if they did. This also that movestart events should be able to be triggered programmatically - although that would require adding a bunch of properties to the event (startX, startY, etc), as subsequent events rely on the initial properties.